v0.3.0 Remove prototype, add placeholder project files
Remove the old prototype, only keeping some reusable elements.
Add placeholder files to keep the directory structure
Old prototype will be deleted except for the PasswordService, SessionService and their corresponding tests.

UserRepo , UserService and User model will also be deleted, although a lot of their functionality could be reused when we implement login. I am deleting them because the User model and subsequent classes will be changed and it is better to write it fresh.

Prototype config  (application.properties) will remain though, as well as the logging system. They don't need to be modified (at the movement) an will be used, no point re-implementing.
SQL scripts and database tables will remain and will be modified once we reach that step