v0.9.64 Seperate test and production transactions, Refactor order status
Add a new environment variable, TREECREATE_ENV. It describes the environment type the application is deployed in. Default is development. If it is set to 'production', the transaction IDs will be marked with `order-`. Otherwise they are prefixed with `test-`.

Make the order update (that happens whenever the orders are queried) be asynchronous and happen after the return (next statement is return so the returned values should not be affected).. This will speed up the loading time in the orders tab.

Add a mapping for fetching the environment type the app is deployed in